+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0.a2.dev0] - 2026-08-06
+
+- Model: `gpytorch-qr==0.8.0`
+- Dataset:
+  - `heavyedge/profiles:v1.0.0rc4`
+  - `heavyedge/shape-features:v1.0.0a4`
+
+### Changed
+
+- GPR now uses train data + validation data for final training.
+
 ## [1.0.0.a2] - 2026-08-06
 
 - Model: `gpytorch-qr==0.8.0`